feat: focus style, removed arch slider
