+++ slider/src/components/horizontal/OldHorizontalSlider.vue
@@ -1,0 +1,460 @@
+<template>
+  <div
+    :id="id"
+    ref="wrap"
+    class="slider"
+    :style="calculateHeight"
+    @click="wrapClick"
+  >
+    <Track
+      class="slider__label-track"
+      :value="val"
+      :min-value="minValue"
+      :max-value="maxValue"
+    >
+      <Tooltip class="track__value-indicator">
+        {{ value }}
+      </Tooltip>
+    </Track>
+    <Track
+      class="track_tmp"
+      :value="val"
+      :min-value="minValue"
+      :max-value="maxValue"
+    >
+      <Thumb
+        :on-mouse-down="moveStart"
+        :on-touch-start="moveStart"
+      />
+      <template #background>
+        <div
+          ref="process"
+          :style="`width: ${position}px`"
+          class="slider"
+        />
+        <div class="slider__rail" />
+      </template>
+    </Track>
+    <div
+      ref="elem"
+      class="slider-bar"
+      style="background-color: red"
+    ></div>
+    <div class="slider__label-container">
+      <div class="slider__range-label slider__range-label__min">
+        {{ minValue }}
+      </div>
+      <div class="slider__range-label slider__range-label__max">
+        {{ maxValue }}
+      </div>
+    </div>
+  </div>
+</template>
+<script>
+import Track from '@/components/horizontal/Track'
+import Thumb from '@/components/Thumb'
+import Tooltip from '@/components/horizontal/Tooltip'
+
+export default {
+  name: 'HorizontalSlider',
+  components: { Thumb, Track, Tooltip },
+  props: {
+    data: {
+      type: Array,
+      default: null,
+    },
+    id: {
+      type: String,
+      default: 'wrap',
+    },
+    // Keep
+    setValue: {
+      type: Function,
+      default: () => {},
+    },
+    value: {
+      type: Number,
+      default: 0,
+    },
+    minValue: {
+      type: Number,
+      default: 0,
+    },
+    maxValue: {
+      type: Number,
+      default: 100,
+    },
+  },
+  data() {
+    return {
+      flag: false,
+      size: 0,
+      currentValue: 0,
+      currentSlider: 0,
+      isComponentExists: true,
+      interval: 1,
+      lazy: false,
+      realTime: false,
+    }
+  },
+  computed: {
+    val: {
+      get() {
+        return this.data ? this.data[this.currentValue] : this.currentValue
+      },
+      set(val) {
+        if (this.data) {
+          let index = this.data.indexOf(val)
+          if (index > -1) {
+            this.currentValue = index
+          }
+        } else {
+          this.currentValue = val
+        }
+      },
+    },
+    currentIndex() {
+      return (this.currentValue - this.minimum) / this.spacing
+    },
+    minimum() {
+      return this.data ? 0 : this.minValue
+    },
+    maximum() {
+      return this.data ? this.data.length - 1 : this.maxValue
+    },
+    multiple() {
+      let decimals = `${this.interval}`.split('.')[1]
+      return decimals ? Math.pow(10, decimals.length) : 1
+    },
+    spacing() {
+      return this.data ? 1 : this.interval
+    },
+    total() {
+      if (this.data) {
+        return this.data.length - 1
+      } else if (
+        Math.floor((this.maximum - this.minimum) * this.multiple) %
+          (this.interval * this.multiple) !==
+        0
+      ) {
+        this.printError(
+          '[VueSlideBar error]: Prop[interval] is illegal, Please make sure that the interval can be divisible',
+        )
+      }
+      return (this.maximum - this.minimum) / this.interval
+    },
+    gap() {
+      return this.size / this.total
+    },
+    position() {
+      return ((this.currentValue - this.minimum) / this.spacing) * this.gap
+    },
+    limit() {
+      return [0, this.size]
+    },
+    valueLimit() {
+      return [this.minimum, this.maximum]
+    },
+    calculateHeight() {
+      return {}
+    },
+  },
+  watch: {
+    value(val) {
+      this.setVal(val)
+    },
+    maxValue(val) {
+      if (val < this.minValue) {
+        return this.printError(
+          '[VueSlideBar error]: The maximum value can not be less than the minimum value.',
+        )
+      }
+      let resetVal = this.limitValue(this.val)
+      this.setVal(resetVal)
+      this.refresh()
+    },
+    minValue(val) {
+      if (val > this.maxValue) {
+        return this.printError(
+          '[VueSlideBar error]: The minimum value can not be greater than the maximum value.',
+        )
+      }
+      let resetVal = this.limitValue(this.val)
+      this.setVal(resetVal)
+      this.refresh()
+    },
+  },
+  mounted() {
+    this.isComponentExists = true
+    if (typeof window === 'undefined' || typeof document === 'undefined') {
+      return this.printError(
+        '[VueSlideBar error]: window or document is undefined, can not be initialization.',
+      )
+    }
+    this.$nextTick(() => {
+      if (this.isComponentExists) {
+        this.getStaticData()
+        this.setVal(this.limitValue(this.value), 0)
+        this.bindEvents()
+      }
+    })
+  },
+  beforeDestroy() {
+    this.isComponentExists = false
+    this.unbindEvents()
+  },
+  methods: {
+    bindEvents() {
+      document.addEventListener('touchmove', this.moving, { passive: false })
+      document.addEventListener('touchend', this.moveEnd, { passive: false })
+      document.addEventListener('mousemove', this.moving)
+      document.addEventListener('mouseup', this.moveEnd)
+      document.addEventListener('mouseleave', this.moveEnd)
+      window.addEventListener('resize', this.refresh)
+    },
+    unbindEvents() {
+      window.removeEventListener('resize', this.refresh)
+      document.removeEventListener('touchmove', this.moving)
+      document.removeEventListener('touchend', this.moveEnd)
+      document.removeEventListener('mousemove', this.moving)
+      document.removeEventListener('mouseup', this.moveEnd)
+      document.removeEventListener('mouseleave', this.moveEnd)
+    },
+    getPos(e) {
+      this.realTime && this.getStaticData()
+      return e.clientX - this.offset
+    },
+    wrapClick(e) {
+      let pos = this.getPos(e)
+      this.setValueOnPos(pos)
+    },
+    moveStart() {
+      this.flag = true
+      this.$emit('dragStart', this)
+    },
+    moving(e) {
+      if (!this.flag) {
+        return false
+      }
+      e.preventDefault()
+      if (e.targetTouches && e.targetTouches[0]) {
+        e = e.targetTouches[0]
+      }
+      this.setValueOnPos(this.getPos(e), true)
+    },
+    moveEnd() {
+      if (this.flag) {
+        this.$emit('dragEnd', this)
+        if (this.lazy && this.isDiff(this.val, this.value)) {
+          this.syncValue()
+        }
+      } else {
+        return false
+      }
+      this.flag = false
+    },
+    setValueOnPos(pos, isDrag) {
+      let range = this.limit
+      let valueRange = this.valueLimit
+      if (pos >= range[0] && pos <= range[1]) {
+        let v =
+          (Math.round(pos / this.gap) * (this.spacing * this.multiple) +
+            this.minimum * this.multiple) /
+          this.multiple
+        this.setCurrentValue(v, isDrag)
+      } else if (pos < range[0]) {
+        this.setCurrentValue(valueRange[0])
+        if (this.currentSlider === 1) {
+          this.currentSlider = 0
+        }
+      } else {
+        this.setCurrentValue(valueRange[1])
+        if (this.currentSlider === 0) {
+          this.currentSlider = 1
+        }
+      }
+    },
+    isDiff(a, b) {
+      if (
+        Object.prototype.toString.call(a) !== Object.prototype.toString.call(b)
+      ) {
+        return true
+      } else if (Array.isArray(a) && a.length === b.length) {
+        return a.some((v, i) => v !== b[i])
+      }
+      return a !== b
+    },
+    setCurrentValue(val) {
+      if (val < this.minimum || val > this.maximum) {
+        return false
+      }
+      if (this.isDiff(this.currentValue, val)) {
+        this.currentValue = val
+        if (!this.lazy || !this.flag) {
+          this.syncValue()
+        }
+      }
+    },
+    setVal(val) {
+      if (this.isDiff(this.val, val)) {
+        let resetVal = this.limitValue(val)
+        this.val = resetVal
+        this.syncValue()
+      }
+    },
+    limitValue(val) {
+      if (this.data) {
+        return val
+      }
+      const inRange = (v) => {
+        if (v < this.minValue) {
+          this.printError(
+            `[VueSlideBar warn]: The value of the slider is ${val}, the minimum value is ${this.minValue}, the value of this slider can not be less than the minimum value`,
+          )
+          return this.minValue
+        } else if (v > this.maxValue) {
+          this.printError(
+            `[VueSlideBar warn]: The value of the slider is ${val}, the maximum value is ${this.maxValue}, the value of this slider can not be greater than the maximum value`,
+          )
+          return this.maxValue
+        }
+        return v
+      }
+      return inRange(val)
+    },
+    syncValue() {
+      let val = this.val
+      if (this.range) {
+        this.$emit('callbackRange', this.range[this.currentIndex])
+      }
+      this.setValue(val)
+    },
+    getStaticData() {
+      if (this.$refs.elem) {
+        this.size = this.$refs.elem.offsetWidth
+        this.offset = this.$refs.elem.getBoundingClientRect().left
+      }
+    },
+    refresh() {
+      if (this.$refs.elem) {
+        this.getStaticData()
+        this.setPosition()
+      }
+    },
+    printError(msg) {
+      console.error(msg)
+    },
+  },
+}
+</script>
+
+<style scoped lang="scss">
+@import '../../styles';
+$arrowHeight: 5px;
+$rail-height: 6px;
+$thumb-radius: 13px;
+$padding-label: 5px 10px;
+$margin-top: 3px;
+
+.slider {
+  position: relative;
+  box-sizing: border-box;
+  user-select: none;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  overflow: hidden;
+  padding: 0px 20px; //tmp
+
+  & .slider__label-track {
+    @include transition(opacity);
+  }
+  &:not(:hover) .slider__label-track {
+    opacity: 0;
+  }
+  &:hover .slider__label-track {
+    opacity: 1;
+  }
+}
+
+.track_tmp {
+  cursor: pointer;
+}
+
+.slider-bar {
+  position: relative;
+  cursor: pointer;
+  height: #{2 * $thumb-radius};
+  margin: $margin-top 0px;
+}
+
+.slider__rail {
+  position: absolute;
+  width: 100%;
+  height: $rail-height;
+  z-index: -2;
+  background-color: $color-grey;
+  border-radius: calc($rail-height / 2);
+  @include centerY;
+}
+.slider {
+  position: absolute;
+  background-color: $color-teal;
+  transition: all 0s;
+  z-index: -1;
+  width: 0;
+  left: 0;
+  will-change: width;
+  border-radius: 15px 0 0 15px;
+  height: $rail-height;
+  @include transition(width);
+  @include centerY;
+}
+
+.slider__thumb__container {
+  position: absolute;
+  cursor: pointer;
+  z-index: 3;
+  left: 0;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  @include transition(left);
+  &:hover .slider__range-label__top {
+    opacity: 1;
+  }
+}
+.slider__thumb {
+  position: absolute;
+  cursor: pointer;
+  z-index: 3;
+  left: 0;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  @include transition(left);
+  &:hover .slider__range-label__top {
+    opacity: 1;
+  }
+}
+
+.slider__label-container {
+  height: 1em; // todo
+  line-height: 2em;
+  position: relative;
+}
+
+.slider__range-label {
+  //padding: $padding-label;
+  position: absolute;
+  color: $color-text-secondary;
+  top: 50%;
+  @include typography-label;
+  &__min {
+    transform: translate(-50%, -50%);
+    left: 0;
+  }
+  &__max {
+    transform: translate(50%, -50%);
+    right: 0;
+  }
+}
+</style>
